Moved MaintenanceModeModal to basic components. Improved styling. Added Services Statuses for allowed user on Maintenance Mode Page.
